Bring over Key Points from version 5.2
Having the key points i.e. summary before the challenges is helpful to the learners I believe. I have made a small change to the key points by replacing the word "snapshot" which is not defined anywhere else with "revisions", which is defined and reused a few times. If want to introduce to snapshots, perhaps better placed in the collaboration section. Finally some formatting of my markdown needs to occur as am not sure what to put in {.xxxxxx} for key points