Add exercise to explore commonly used licenses
This exercise is designed to encourage the learners to be more aware of the sorts of software licenses they have already accepted. Even amongst the tools used in the Software Carpentry workshops, there are several different licenses used. Students may discuss with others and/or contribute to a list in the EtherPad.