Removes reference to push -u, and intermediate git lesson
The -u flag isn't explained, other than to ask the user to ignore it.
It makes more sense not to mention it at all - it only adds confusion.

The referenced intermediate lesson isn't linked to, and until it is a
lesson on the Software Carpentry page, it only adds further confusion.