Added understanding HEAD as an objective and text to support it.
I grepped HEAD in previous sections, and this is it's first appearance.  To
my mind, it should be explicitly introduced, because it is a source of
some confusion.  I added the learning objective and the text to support
it to the top of the lesson.  I tried to merge without seam into what
was already there.