Recommend against nest git repos
Changed to neutral wording, rather than "bad thing" and added more explicit recommendation against nested repos.

Added `ls -a` to the question to make it easier for participants unfamiliar with the terminal to deduce `rm -rf .git` as the answer (as it will be displayed in the terminal).

Example given below on how to demonstrate how to track within nested directories. I note that inclusion of the later content (ignore) may not be suitable here.