Changes to text editor configuration
Updated config to recommend github account as user.name

Added Scratch (Linux) as a text editor for Linux users, this is the default editor for Elementary OS

Noted that Vim is the default editor, previous Vim users have asked helpers to configure it instead of nano (which we usually demonstrate in the workshop). Included instructions for novices to exit Vim if configuration unintentionally left on default.