04-changes.md: make Directories example more explicit
Directories are an important concept in git, and this exercise is a useful one,
but it can be improved.

This commit improves the example in two ways.
Firstly, we now use the name "spaceships" for the directory instead of simply
"directory"; this makes it clear to learners which part of the command is syntax
and which is a name.
Secondly, we add an extra code block that explicitly creates files and adds a
directory.  The final `git status` command shows the user that the two new files
have been added.

This was done as a contribution for Instructor Training Checkout.