Update the objectives in episode 1
Updated the second objective in episode 1 to solve issue #519. I like the original suggestion in the issue since it's more consistent with the first objective, and sounds less complex to me. At this point the student really only needs to very broadly understand what a version control system does, not so much what operations are performed.