Indicate who makes what changes
Having the collaborator make the first change puts the responsibility to resolve the conflict with the owner, which makes more sense. I have often seen the reverse scenario during workshops, and then one has to explain that the collaborator needs to contact the owner to discuss how to resolve it, as they are not the original author. It makes more sense to have the original author/owner make the decision how to resolve the conflict.