Use more inclusive beverage example
This is a very minor change, but I do think it's worth considering. In the "A Typical Work Session" exercise in 09-conflict, one of the typical programmer actions presented is `*Celebrate* your success with beer(s)`. While beer is definitely the celebratory beverage of choice for many of us, there are certain communities and parts of the world where alcoholic beverages are viewed more sternly or in a negative light. Also, there are many potential learners who have not yet reached the legal drinking age in their region.

I believe that tweaking this language could make the exercise a little bit more inclusive, relatable, and fun. While a simple adaptation would be "non-alcoholic beer(s)", I think it is more fun to say "fancy beverage(s)", since this conjures in each mind some favorite drink which brings the learner joy. I welcome anyone to comment below with any alternative suggestions!