Explain the use of GitHub personal access tokens in the remotes episode
Github has deprecated password authentication from the command line
and will disable it in August 2021: https://github.blog/2020-12-15-token-authentication-requirements-for-git-operations

This change instead explains in the "Working with remotes" episode
how to generate a personal access token which is the natural replacement
method of password authentication.

It allows adds some instructor notes on how to handle tokens in live coding.