Update alt-text for RStudio screenshot
Note that this particular alt-text is a bit long. It can be shortened if you want to describe the files that are added by default in the text.

This contribution is part of the Core Team accessibility hackathon to add alt-text to images across our lessons. Please feel free to improve upon these suggestions as desired!