Renamed references for "master" branch to "main" branch
Use of a "master" branch is generally discouraged as culturally insensitive and the most recent versions of git default to a "main" branch instead.

Even if some Carpentries users are using older versions of git that still default to a "master" branch, I think it would be best for the language on the website to reflect the more culturally sensitive use.