Update 14-supplemental-rstudio.md
I propose changing 'master' to 'main' one time in the text for this page. Also, two of the figures will need to be updated: RStudio_screenshot_review.png and RStudio_screenshot_viewhistory.png contain references to the master branch that will need to be changed to main branch to be consistent with the text.