Update guide.md
Updated the `git branch [...]` command to reflect GitHub's stated CLI command, `git branch -M main`.  Also, this command will only work if there are commits in the branches, otherwise it returns a fatal error because the branches aren't found.  When changing the main branch name in a newly created (and thus empty) repository, the command is `git checkout -b main`