Incorrect number of commits before head for shown output.
As I followed along, the $ git diff HEAD~3 is back one commit further than my history. It is possible if a previous aside is followed that this might not generate an error about that version not existing, however, the output shown indicates no such commit. So the command is inconsistent with the output shown. A simple fix is to replace the "3" with a "2"